+++ lib/src/commands/commands.dart
@@ -1,3 +1,4 @@
 export 'auth_command.dart';
 export 'init_command.dart';
 export 'feature_command.dart';
+export 'service_command.dart';